refactor(integration): extract detector subprocess execution into health_checker_tools.go
health_checker.go was ~1k lines. Move the cohesive subprocess-execution block —
runFFprobeWithArgs, runHandBrakeWithArgs, buildMediaInfoArgs,
runCommandWithTimeout, validateMediaInfoOutput, hasValidMediaTrack, runMediaInfo —
verbatim into health_checker_tools.go (same package). Pure code move: no logic,
signature, or behavior change. health_checker.go drops 1021 -> 781 lines; the
public Check/CheckWithConfig entry points and classification logic stay put.

Test plan:
- go build ./... ; golangci-lint ./internal/integration/... 0 issues
- go test ./internal/integration/ pass
